[ALPS06907273] ACK: Revert android12-5.10-2022-02 R1
Revert "[ALPS06907273] ACK: Merge
        android12-5.10-2022-02 R1 into alps-mp-s0.mp1"
This reverts commit 67d967bc27511d5a971e506667b749539ee93dd0.

MTK-Commit-Id: e71082dbeb1be82b53e0cc68b5108b0874da839e

Feature: Kernel SI Operation
CR-Id: ALPS06907273

+++ drivers/scsi/ufs/ufs-exynos.h
@@ -197,7 +197,7 @@ struct exynos_ufs {
 	u32 pclk_div;
 	u32 pclk_avail_min;
 	u32 pclk_avail_max;
-	unsigned long mclk_rate;
+	u32 mclk_rate;
 	int avail_ln_rx;
 	int avail_ln_tx;
 	int rx_sel_idx;
